Change logic and rename ProposalTypeSelect
